# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iterra API",
    "description": "AI Content Strategy Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/trends": {
      "get": {
        "tags": [
          "trends"
        ],
        "summary": "Get Trends",
        "operationId": "get_trends_api_v1_trends_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trends/refresh": {
      "post": {
        "tags": [
          "trends"
        ],
        "summary": "Refresh Trends",
        "operationId": "refresh_trends_api_v1_trends_refresh_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/storage/status": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "Storage Status",
        "operationId": "storage_status_api_v1_storage_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storage/export": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "Export Data",
        "description": "Lists all Iterra files in user's Drive. Files are in Drive \u2014 user downloads directly.",
        "operationId": "export_data_api_v1_storage_export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageExportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storage/data": {
      "delete": {
        "tags": [
          "storage"
        ],
        "summary": "Delete All Data",
        "description": "GDPR: Deletes all Iterra-created files from user's Drive.\nClears drive_file_id and drive_analysis_file_id from our DB.\nDoes NOT delete the user account or non-Iterra Drive files.",
        "operationId": "delete_all_data_api_v1_storage_data_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteDataResponse": {
        "properties": {
          "deleted_files": {
            "type": "integer",
            "title": "Deleted Files"
          },
          "db_records_cleared": {
            "type": "boolean",
            "title": "Db Records Cleared"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "deleted_files",
          "db_records_cleared",
          "message"
        ],
        "title": "DeleteDataResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StorageExportResponse": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/StorageFileInfo"
            },
            "type": "array",
            "title": "Files"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Files listed from your Google Drive. Download them directly from Drive."
          }
        },
        "type": "object",
        "required": [
          "files",
          "total"
        ],
        "title": "StorageExportResponse"
      },
      "StorageFileInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "size": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "modified_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "StorageFileInfo"
      },
      "StorageStatus": {
        "properties": {
          "connected": {
            "type": "boolean",
            "title": "Connected"
          },
          "iterra_folder_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iterra Folder Id"
          },
          "drafts_folder_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drafts Folder Id"
          },
          "scraped_posts_file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scraped Posts File Id"
          },
          "brand_analysis_file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Analysis File Id"
          }
        },
        "type": "object",
        "required": [
          "connected"
        ],
        "title": "StorageStatus"
      },
...
      "TrendItemResponse": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "raw_score": {
            "type": "integer",
            "title": "Raw Score"
          },
          "niche_relevance": {
            "type": "number",
            "title": "Niche Relevance"
          },
          "related_queries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Related Queries"
          },
          "content_angle": {
            "type": "string",
            "title": "Content Angle"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "raw_score",
          "niche_relevance",
          "related_queries",
          "content_angle"
        ],
        "title": "TrendItemResponse"
      },
      "TrendResponse": {
        "properties": {
          "niche": {
            "type": "string",
            "title": "Niche"
          },
          "trends": {
            "items": {
              "$ref": "#/components/schemas/TrendItemResponse"
            },
            "type": "array",
            "title": "Trends"
          },
          "top_pick": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrendItemResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "cache_age_minutes": {
            "type": "integer",
            "title": "Cache Age Minutes"
          }
        },
        "type": "object",
        "required": [
          "niche",
          "trends",
          "fetched_at",
          "cache_age_minutes"
        ],
        "title": "TrendResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}